ARR30
git-svn-id: svn://svn.code.sf.net/p/rosecheckers/code/trunk@159 a2ac0d72-06d8-44ab-9e58-0f2374626789

+++ c_rules/ncce/ARR_NCCE.c
@@ -71,7 +71,7 @@ void ARR02(void) {
 
 int *table2 = NULL;
 
-int ARR30_NCCE_insert_in_table(size_t pos, int value){
+int ARR30_NCCE_insert_in_table(int pos, int value){
 enum { TABLESIZE = 100 };
   if (!table2) {
     table2 = (int *)malloc(sizeof(int) * TABLESIZE);
